Some cleanup
* clean Ogre imports
  - remove #include Ogre.h, add precise #includes
* rename CeGuiWindow to AbstractWindow 
  - removes issue with similar CeGuiWindow.h/CEGUIWindow.h files
* removed some warnings

git-svn-id: svn://svn.berlios.de/dsa-hl@3160 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/core/src/CoreSubsystem.cpp
@@ -287,7 +287,7 @@ namespace rl
         // Initialise the modules
         Ogre::StringVector modulesList = ConfigurationManager::getSingleton().getModuleList();
 
-        for (int i = 0; i < modulesList.size(); i++)
+        for (size_t i = 0; i < modulesList.size(); i++)
         {
             mRubyInterpreter->executeFile(ContentModule::getInitFile(modulesList[i]));
